Revert "Removing react-select usage by either not using SingleValue or moving some components in package"
