+++ package.json
@@ -51,6 +51,7 @@
     "react-notifications-component": "^4.0.1",
     "react-oidc-context": "^2.1.0",
     "react-paginating": "^1.4.0",
+    "react-select": "^5.7.4",
     "react-simplemde-editor": "^5.0.2",
     "react-textarea-autosize": "^8.5.2",
     "react-transition-group": "^4.4.5",
