+++ packages/components/src/index.ts
@@ -1,8 +1,6 @@
-export * from './AzurePersistentDiskSizeSelectInput';
 export * from './buttons';
 export * from './Clickable';
 export * from './DelayedRender';
-export * from './disk-utils';
 export * from './ErrorBoundary';
 export * from './FocusTrap';
 export * from './hooks/useUniqueId';
@@ -14,7 +12,6 @@ export * from './injectStyle';
 export { Interactive } from './Interactive';
 export * from './Link';
 export * from './Modal';
-export * from './PersistentDiskTypeInput';
 export * from './PopupTrigger';
 export * from './Select';
 export * from './Spinner';
