+++ src/analysis/modals/ComputeModal/AzureComputeModal/AzurePersistentDiskSizeSelectInput.ts
@@ -1,13 +1,20 @@
-import {
-  AzurePersistentDiskSizeSelect,
-  AzurePersistentDiskSizeSelectInputProps,
-  useUniqueId,
-} from '@terra-ui-packages/components';
+import { useUniqueId } from '@terra-ui-packages/components';
 import React from 'react';
 import { div, h, label } from 'react-hyperscript-helpers';
+import { SingleValue } from 'react-select';
+import { IComputeConfig } from 'src/analysis/modal-utils';
 import { computeStyles } from 'src/analysis/modals/modalStyles';
+import { Select } from 'src/components/common';
 import { azureDiskSizes } from 'src/libs/ajax/leonardo/models/disk-models';
 import { defaultAzureDiskSize } from 'src/libs/azure-utils';
+
+export interface AzurePersistentDiskSizeSelectInputProps {
+  persistentDiskSize: number;
+  onChangePersistentDiskSize: (e: SingleValue<number | undefined>) => void;
+  persistentDiskExists: boolean;
+}
+
+const AzurePersistentDiskSizeSelect = Select as typeof Select<IComputeConfig['persistentDiskSize']>;
 
 export const AzurePersistentDiskSizeSelectInput: React.FC<AzurePersistentDiskSizeSelectInputProps> = (
   props: AzurePersistentDiskSizeSelectInputProps
